+++ lib/mongoid/contextual.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+require "mongoid/contextual/map_reduce"
 require "mongoid/contextual/memory"
 require "mongoid/contextual/mongo"