+++ spec/mongoid/errors/no_map_reduce_output_spec.rb
@@ -1,0 +1,29 @@
+require "spec_helper"
+
+describe Mongoid::Errors::NoMapReduceOutput do
+
+  describe "#message" do
+
+    let(:error) do
+      described_class.new(query: {})
+    end
+
+    it "contains the problem in the message" do
+      error.message.should include(
+        "No output location was specified for the map/reduce operation."
+      )
+    end
+
+    it "contains the summary in the message" do
+      error.message.should include(
+        "When executing a map/reduce, you must provide the output location"
+      )
+    end
+
+    it "contains the resolution in the message" do
+      error.message.should include(
+        "Provide the location that the output of the operation"
+      )
+    end
+  end
+end